pyproject.toml
[project]
name = "volnateca-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.136.1,<0.137.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "sqlmodel (>=0.0.38,<0.0.39)",
    "ruff (>=0.15.12,<0.16.0)",
    "mypy (>=1.20.2,<2.0.0)",
    "dishka (>=1.10.1,<2.0.0)",
    "uvicorn (>=0.46.0,<0.47.0)",
    "pydantic-settings (>=2.14.0,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "aiohttp (>=3.13.5,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.ruff]
line-length = 110
